requirementos.txt
matplotlib.pyplot
numpy
pulp